Avoid circular dependency issues by moving prepare_basic_data_query to its own module
