+++ src/pysdmx/api/dc/__init__.py
@@ -4,12 +4,10 @@ from pysdmx.api.dc._api import (
     BasicConnector,
     Connector,
     MaintainableIdentification,
-    prepare_basic_data_query,
 )
 
 __all__ = [
     "BasicConnector",
     "Connector",
     "MaintainableIdentification",
-    "prepare_basic_data_query",
 ]
